重构了项目，为两个模块提取了公共的配置项和插件等信息。
