+++ docs/ksp.md
@@ -1,9 +1,9 @@
 # MyBatis-Flex KSP 介绍及其配置
 
-在纯 Kotlin/JVM 环境或是 Kotlin 和 Java 的混合环境下，也许我们更希望项目在编译时生成的是 Kotlin 代码而非 Java 代码。
+在 Kotlin/JVM 项目中，也许我们更希望项目在编译时生成的 `TableDef` 类是 Kotlin 代码而非 Java 代码。
 如果您的项目**使用 gradle 进行构建**，而您恰好有这样的需求，那么 KSP 将是一个更好的选择。
 
-KSP 生成的代码结构与 APT 近乎一致，并且在 Kotlin 的角度充分考虑了兼容性。如有需要，您可以放心大胆地把先有的 APT 更换成 KSP，然后根据此文档做进一步地了解和配置。
+MyBatis-Flex KSP 生成的代码结构与 MyBatis-Flex APT 近乎一致，并且在 Kotlin 的角度充分考虑了兼容性。如有需要，您可以放心大胆地把先有的 APT 更换成 KSP，然后根据此文档做进一步地了解和配置。
 
 ## 什么是 KSP ？为什么使用 KSP ？
 
@@ -107,41 +107,36 @@ data class Account(
 @file:Suppress("RedundantVisibilityModifier", "MemberVisibilityCanBePrivate", "unused", "RemoveRedundantBackticks")
 
 public object AccountTableDef : TableDef("", "tb_account") {
-    /**
-     * 主键
-     * @return id
-     */
-    public val ID: QueryColumn = QueryColumn(this, "id")
+    @JvmField
+    public val ID: QueryColumn = QueryColumn(this,  "ID")
 
-    public val USER_NAME: QueryColumn = QueryColumn(this, "user_name")
+    @JvmField
+    public val USER_NAME: QueryColumn = QueryColumn(this,  "USER_NAME")
 
-    public val AGE: QueryColumn = QueryColumn(this, "age")
+    @JvmField
+    public val AGE: QueryColumn = QueryColumn(this,  "AGE")
 
-    public val BIRTHDAY: QueryColumn = QueryColumn(this, "birthday")
+    @JvmField
+    public val BIRTHDAY: QueryColumn = QueryColumn(this,  "BIRTHDAY")
 
+    @JvmField
     public val ALL_COLUMNS: QueryColumn = QueryColumn(this, "*")
 
-    public val DEFAULT_COLUMNS: List<QueryColumn> = listOf(`ID`, `USER_NAME`, `AGE`, `BIRTHDAY`)
+    @JvmField
+    public val DEFAULT_COLUMNS: List<QueryColumn> = listOf(`ID`,`USER_NAME`,`AGE`,`BIRTHDAY`)
 
     @JvmField
     public val ACCOUNT: AccountTableDef = AccountTableDef
-
-    /**
-     * A constructor used to mock objects for code compatibility. It returns itself on each call.
-     */
-    @Suppress("NOTHING_TO_INLINE")
-    public inline operator fun invoke() = this
 }
 ```
 
-对比于 APT 生成的产物，我们可以很明显地发现，
-KSP 的产物中，`AccountTableDef` 是一个 Kotlin 对象，
-而 APT 的产物中，`AccountTableDef` 是一个普通的 Java 类。
+对比于 APT 生成的产物，我们可以很明显地发现以下区别：
 
-`AccountTableDef` 还拥有一个 `invoke` 方法，和一个 `ACCOUNT` 属性。
-这个方法和属性均是为了兼容 APT 而生， 以使得我们不需要做太多的改动即可从 APT 迁移到 KSP。
+1. KSP 的产物中，`AccountTableDef` 是一个 Kotlin 对象，而 APT 的产物中，`AccountTableDef` 是一个普通的 Java 类。公共构造器被移除了。
 
-此外，`AccountTableDef` 生成的 `DEFAULT_COLUMNS` 的类型由 `Array` 变为了 `List`。
+2. `AccountTableDef` 拥有一个指向自身的 `ACCOUNT` 属性。这个属性是为了兼容 APT 而生，以使得我们不需要做太多的改动即可从 APT 迁移到 KSP。
+
+3. `AccountTableDef` 生成的 `DEFAULT_COLUMNS` 的类型由 `Array` 变为了 `List`。 
 这是为了符合 Kotlin 语言常用习惯而特意做的修改。如果您确实需要一个 `Array` ，您可以在稍后的 KSP 配置部分进行配置。
 
 ## KSP 相关配置
@@ -164,27 +159,60 @@ ksp {
 这里我们将两个参数传入至 KSP 中。其中第一个参数 `flex.project.path` 用于告诉 KSP 当前项目根目录的路径。
 
 而第二个参数 `flex.root.project.path` 用于指定您的项目最外层的项目所在。
-这个参数的作用是用于 `processor.stopBubbling` 配置以告知 KSP 向上级合并配置时应该到哪里结束合并。
-因此如果您没有合并配置的需求，您可以不写此参数。
+这个参数的作用是用于 `processor.stopBubbling` 配置以告知 KSP 向上级合并配置时应该到哪里结束合并，
+因此如果您没有合并配置的需求，您不必写此参数。
 
 ### KSP 配置与 APT 配置的异同点及其相关说明
 
-KSP 几乎完全兼容 APT 的配置。而且我们针对一些配置做了进一步地改善和增强。接下来我们来一起看看 KSP 究竟做了哪些优化。
+KSP 几乎完全兼容 APT 的配置。此外，我们针对一些配置还做了进一步地改善和增强。
 
-1. **processor.enable**
+接下来我们来一起看看对于 KSP 而言不支持和做了改动的配置。对于下面未曾提到的 APT 配置，KSP 均完全支持。
 
-此配置用于指定是否启用 KSP 。我们的 KSP 对此配置实现了完全的支持。
-但我们不建议您这样做，而是更推荐您将此配置项写入 `build.gradle` 文件中，或通过其他方法来决定是否关闭 KSP。
+#### processor.enable
 
-您应该在 `build.gradle` 文件中进行如下配置，以决定是否启用 KSP：
+    此配置用于指定是否启用 KSP 。我们的 KSP 对此配置实现了完全的支持。
+    但我们不建议您这样做，而是更推荐您**直接将依赖项中我们的 KSP 行注释掉它或者直接删除它**。这样做最为简单省事，也能更好地节省性能开销。
+    
+    如果您不想这么做，那么我们推荐您在 `build.gradle` 文件中进行如下配置，以决定是否启用 KSP：
+    
+    【Groovy】/【Kotlin】
+    ```
+    ksp {
+        arg("flex.ksp.enable", "false")
+    }
+    ```
+    
+    我们知道，如果我们不想让 KSP 为 flex 生成代码，那么 `mybatis-flex.config` 配置文件中的内容对于 KSP 来说也是毫无意义的。
+    在 `build.gradle` 文件中进行配置，可以使得 KSP 在一开始时便知道不需要为 flex 生成代码，因此可以避免很多诸如读取配置文件等不必要的操作以节省性能开销。
 
-【Groovy】/【Kotlin】
-```
-ksp {
-    arg("flex.ksp.enable", "false")
-}
-```
+#### processor.stopBubbling
 
-我们知道，如果我们不想让 KSP 为 flex 生成代码，那么 `mybatis-flex.config` 配置文件中的内容对于 KSP 来说也是毫无意义的。
-在 `build.gradle` 文件中进行配置，可以使得 KSP 在一开始便知道不需要为 flex 生成代码，因此可以节省很多诸如读取配置文件等不必要的性能开销。
+    此配置用于指定是否停止向上级合并配置。KSP 对此完全支持。
+    
+    对于所有 `mybatis-flex.config` 中的配置，在向上合并时，对于重复的键，KSP 会采取以下策略：
+    
+    1. 对于仅能指定一个值的键，KSP 会覆盖原有的值。
+    2. 对于能够指定多个值的键，例如 `processor.tableDef.ignoreEntitySuffixes`，KSP 会将原有的值与新指定的值进行合并。即这些键的值最终都会成为处理的一部分。
+
+#### processor.genPath
+
+    不支持。
+
+#### processor.allInTables.className
+
+    KSP 对此生成的是单例对象，即 `object` ，而不是普通类。
+
+#### processor.allInTables.enable, processor.allInTables.package
+
+    如果您指定 KSP 为您生成 `Tables` 单例（即 `processor.allInTables.enable` 的值为 `true`），而又不指定其包名或指定了包名但包名非法，
+    那么 KSP 会向您发出警告，并且不会生成 `Tables` 类。
+
+#### processor.tableDef.propertiesNameStyle
+
+    除了原本的四种风格外，KSP 还支持原始风格。这种风格可以使得 KSP 生成的字段对比于实体类不做任何改变。
+    假设您在实体类中定义的其中一个属性的名字为 `<!-- 非法的属性名 --!>`，那么 KSP 生成的对应字段的名字也是 `<!-- 非法的属性名 --!>`。
+
+### KSP 特有配置
+
+KSP 在兼容 APT 配置的基础上，还增添了一些自己独有的配置。下面我们来一起看一看。
