+++ mybatis-flex-kotlin-ksp/src/main/kotlin/com/mybatisflex/kotlin/internal/util/Utils.kt
@@ -280,6 +280,7 @@ val KSClassDeclaration.mapperPackageName: String
  * ```
  * @author CloudPlayer
  */
+@Suppress("unused")
 fun invokeFunction(): FunSpec.Builder {
     val func = FunSpec.builder("invoke")
     func.modifiers += KModifier.OPERATOR
