--- mybatis-flex-kotlin-ksp/src/test/kotlin/Test.kt
@@ -1,8 +1,0 @@
-import org.junit.jupiter.api.Test
-
-class Test {
-    @Test
-    fun test() {
-        println(System.getProperty("user.dir"))
-    }
-}
